# improve further mergePolicies to return faster