The main benefits include

- Single policy inputs for typical cases; no need to copy the policy.
- actionStatementIndex() should only allocate the map when necessary, 
  avoiding it for Deny-only cases and cases with only wildcard actions.
- There is no requirement to clone() statements; we are not planning 
  to modify the resources per statement.

```
benchmark                                                old ns/op     new ns/op     delta
BenchmarkMergePolicies/10Policies_1Stmt_NoDups-8         10099         3993          -60.46%
BenchmarkMergePolicies/10Policies_1Stmt_HalfDups-8       9597          3420          -64.36%
BenchmarkMergePolicies/100Policies_1Stmt_NoDups-8        90724         35601         -60.76%
BenchmarkMergePolicies/100Policies_1Stmt_HalfDups-8      93520         30534         -67.35%
BenchmarkMergePolicies/1000Policies_1Stmt_NoDups-8       1069535       375023        -64.94%
BenchmarkMergePolicies/1000Policies_1Stmt_HalfDups-8     985113        312909        -68.24%

benchmark                                                old allocs     new allocs     delta
BenchmarkMergePolicies/10Policies_1Stmt_NoDups-8         99             28             -71.72%
BenchmarkMergePolicies/10Policies_1Stmt_HalfDups-8       92             21             -77.17%
BenchmarkMergePolicies/100Policies_1Stmt_NoDups-8        1015           208            -79.51%
BenchmarkMergePolicies/100Policies_1Stmt_HalfDups-8      965            158            -83.63%
BenchmarkMergePolicies/1000Policies_1Stmt_NoDups-8       10024          2012           -79.93%
BenchmarkMergePolicies/1000Policies_1Stmt_HalfDups-8     9522           1510           -84.14%

benchmark                                                old bytes     new bytes     delta
BenchmarkMergePolicies/10Policies_1Stmt_NoDups-8         11080         2576          -76.75%
BenchmarkMergePolicies/10Policies_1Stmt_HalfDups-8       10648         2144          -79.86%
BenchmarkMergePolicies/100Policies_1Stmt_NoDups-8        110642        22880         -79.32%
BenchmarkMergePolicies/100Policies_1Stmt_HalfDups-8      107554        19792         -81.60%
BenchmarkMergePolicies/1000Policies_1Stmt_NoDups-8       1222377       281495        -76.97%
BenchmarkMergePolicies/1000Policies_1Stmt_HalfDups-8     1169183       228302        -80.47%
```